Split Typed `type` property into `resultingType` and `expressionType`
The `resultingType` is used as the resulting type in the parent
expression.

Note that this type DOES NOT need to be equal to the type of the
expression itself. For instance, in the presence of multiple
values

    (+ 1 (values 1 2))

(values 1 2) will have type `(values 1 2)`, however the type number
is returned to the parent expression.

And the other way around:

  (lambda (x) x)

x as an expression has type `a`, but `(values a)` is returned to the
parent function.