Refactor type constraints
We need the ability to introduce temporary type variables and
constraints between those. So looking at constraints as a constraint
on expressions' types is not enough anymore.

The expression is kept for reporting purposes in the `source` field
now.

Some higher level constructors are exposed to keep the ability to
define constraints at at level, but they are defined as low-level type
constraints only.